dump analysis_parameters only first time

+++ cdci_data_analysis/analysis/products.py
@@ -1,5 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
+import os
 from builtins import (bytes, str, open, super, range,
                       zip, round, input, int, pow, object, map, zip)
 
@@ -84,12 +85,19 @@ class QueryOutput(object):
     def set_api_code(self, query_dict, url=None):
         self.prod_dictionary['api_code'] = DispatcherAPI.set_api_code(query_dict, url=url)
 
-    def dump_analysis_parameters(self,work_dir,query_dict):
-        file_path=FilePath(file_dir=work_dir, file_name='analysis_parameters.json')
-        with open(file_path.path, 'w')  as outfile:
-            _dict = {k: v for k, v in query_dict.items() if v is not None}
-            my_json_str = json.dumps(_dict, indent=4, sort_keys=True)
-            outfile.write(u'%s' % my_json_str) # why is this formatting necessary?
+    def dump_analysis_parameters(self, work_dir, query_dict):
+        file_path = FilePath(file_dir=work_dir, file_name='analysis_parameters.json')
+        if not os.path.exists(file_path.path):
+            with open(file_path.path, 'w')  as outfile:
+                _dict = {k: v for k, v in query_dict.items() if v is not None}
+                my_json_str = json.dumps(_dict, indent=4, sort_keys=True)
+                # TODO further test if formatting is really needed
+                outfile.write(u'%s' % my_json_str)
+        else:
+            with open(file_path.path) as outfile:
+                dict_analysis_parameters = json.load(outfile)
+            logger.info("analysis_parameters.json file already present, and originally used for the job_id calculation: %s",
+                         json.dumps(dict_analysis_parameters, sort_keys=True, indent=4))
 
     def set_products(self, keys, values):
         for k, v in zip(keys, values):
